+++ src/common.h
@@ -5,9 +5,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#define CORNAMUSA_VERSION "1.16.3"
+#define CORNAMUSA_VERSION "1.17.0"
 #define CORNAMUSA_VERSION_MAJOR 1
-#define CORNAMUSA_VERSION_MINOR 16
-#define CORNAMUSA_VERSION_PATCH 3
+#define CORNAMUSA_VERSION_MINOR 17
+#define CORNAMUSA_VERSION_PATCH 0
 
 #endif /* CORNAMUSA_COMMON_H */